Connected the GUI frames
Server.java, Client.java are not final (current code for testing purposes).
Need to create MessagePanel format instead of textArea. ChannelList, UserList Panels contents are generated through code after logging in and entering the mainframe based on the current database (final decision if it is in buttons format for each channel and user in these panels).